Fixed bugs and imported front side

+++ src/main/java/com/marketplace/repository/ProductsRepository.java
@@ -7,7 +7,8 @@ import org.springframework.stereotype.Repository;
 import java.util.List;
 
 @Repository
-public interface ProductsRepository extends JpaRepository<Product, Integer> {
+public interface ProductsRepository extends JpaRepository<Product, Long> {
     List<Product> getAllBy();
+
     Product save(Product product);
 }
